Replaced `UdpClient`/`UdpServer` with `UdpSocket`
